backend/requirements.txt
Flask==2.0.2
google-cloud-texttospeech==2.6.0
google-cloud-speech==2.1.0
translate-api==4.9.5